+++ packages/noodl-editor/src/editor/src/views/panels/VersionControlPanel/VersionControlPanel.tsx
@@ -33,30 +33,12 @@ import { History } from './components/History';
 import { LocalChanges } from './components/LocalChanges';
 import { MergeConflicts } from './components/MergeConflicts';
 import { useVersionControlContext, VersionControlProvider } from './context';
+import { convertGitRemoteUrlToRepoUrl } from './github';
 
 enum ViewState {
   Default,
   Branches,
   BranchMerge
-}
-
-function convertGitRemoteUrlToRepoUrl(gitRemoteUrl) {
-  // Remove the .git extension if present
-  gitRemoteUrl = gitRemoteUrl.replace(/\.git$/, '');
-
-  // Extract the repository name and owner from the URL
-  const regex = /^(?:https?:\/\/)?github\.com\/([^/]+)\/([^/]+)$/;
-  const match = gitRemoteUrl.match(regex);
-
-  if (match) {
-    const owner = match[1];
-    const repo = match[2];
-    // Construct the GitHub repository URL
-    const repoUrl = `https://github.com/${owner}/${repo}`;
-    return repoUrl;
-  } else {
-    throw new Error('Invalid GitHub Git remote URL');
-  }
 }
 
 function BaseVersionControlPanel() {
